+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Improvements:
 * [OLMIS-3326:](https://openlmis.atlassian.net/browse/OLMIS-3326) The notificationService will now wait for the loadingModalService to close before displaying any notification
 * [OLMIS-3905:](https://openlmis.atlassian.net/browse/OLMIS-3905) Add min-width to select dropdown
 * [OLMIS-3908:](https://openlmis.atlassian.net/browse/OLMIS-3908) Reduced vertical space in page header
+* [OLMIS-4195:](https://openlmis.atlassian.net/browse/OLMIS-4195) Added empty-row directive
 
 Bug fixes:
 * [OLMIS-3806:](https://openlmis.atlassian.net/browse/OLMIS-3806): Fixed date picker to not change selected date